pyproject.toml
[tool.poetry]
name = "contactbook"
version = "0.1.0"
description = ""
authors = ["jslee <dlwhdtjd098@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = ">=4.2,<5.0"
djangorestframework = ">=3.14,<4.0"
django-mysql = "^4.14.0"
drf-spectacular = "^0.27.2"
pillow = "^10.4.0"
mysqlclient = "^2.2.4"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}


[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
pytest-django = "^4.8.0"
pytest-xdist = "^3.6.1"
pytest-cov = "^5.0.0"
factory-boy = "^3.3.0"
pytest-env = "^1.1.3"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.0"
ruff = "^0.5.6"
pyright = "^1.1.374"
black = "^24.8.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
django-debug-toolbar = "^4.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

